Refatorando a entidade Pessoa

+++ #Banco de Dados/04 - CREATE TABLE - Pessoa.sql
@@ -3,7 +3,7 @@ USE [ControleDespesas]
 IF NOT EXISTS (SELECT 1 FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE='BASE TABLE' AND TABLE_NAME='Pessoa') 
 BEGIN
 CREATE TABLE [dbo].[Pessoa](
-	[Id] [int] IDENTITY(1,1) NOT NULL,
+	[Id] [bigint] IDENTITY(1,1) NOT NULL,
 	[IdUsuario] [bigint] NOT NULL,
 	[Nome] [nvarchar](100) NOT NULL,
 	[ImagemPerfil] [text] NOT NULL,
