+++ ControleDespesas_ApiNetCore21_Dapper/ControleDespesas.Domain/Pagamentos/Commands/Output/PagamentoCommandOutput.cs
@@ -7,7 +7,7 @@ namespace ControleDespesas.Domain.Pagamentos.Commands.Output
         public int Id { get; set; }
         public long IdTipoPagamento { get; set; }
         public long IdEmpresa { get; set; }
-        public int IdPessoa { get; set; }
+        public long IdPessoa { get; set; }
         public string Descricao { get; set; }
         public double Valor { get; set; }
         public DateTime DataVencimento { get; set; }
